+++ src/log.ts
@@ -1,0 +1,3 @@
+import * as vscode from "vscode";
+
+export const log = vscode.window.createOutputChannel("Venom", { log: true });
